# CryptoMaui/BackendClient/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CryptoBack",
    "version": "1.0"
  },
  "paths": {
    "/api/Coins/GetInstalledCoins": {
      "get": {
        "tags": [
          "Coins"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Coins/GetPriceHistory": {
      "get": {
        "tags": [
          "Coins"
        ],
        "parameters": [
          {
            "name": "coinId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoinPriceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoinPriceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoinPriceDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Coins/GetInvestments": {
      "get": {
        "tags": [
          "Coins"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TotalInvetmentsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalInvetmentsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalInvetmentsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Coins/AddInvestment": {
      "post": {
        "tags": [
          "Coins"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InvestmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/Login": {
      "post": {
        "tags": [
          "Login"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CoinPriceDto": {
        "required": [
          "date",
          "price"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "price": {
            "type": "number",
            "format": "decimal"
          }
        },
        "additionalProperties": false
      },
      "DcaDto": {
        "type": "object",
        "properties": {
          "coinName": {
            "type": "string",
            "nullable": true
          },
          "dca": {
            "type": "number",
            "format": "decimal"
          }
        },
        "additionalProperties": false
      },
      "InvestmentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "coinAmmount": {
            "type": "number",
            "format": "decimal"
          },
          "investmentValue": {
            "type": "number",
            "format": "decimal"
          },
          "coinName": {
            "type": "string",
            "nullable": true
          },
          "repeatMonthly": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "LoginDto": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "TotalInvetmentsDto": {
        "type": "object",
        "properties": {
          "investments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvestmentDto"
            },
            "nullable": true
          },
          "dcas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DcaDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}